@@ -1,4 +1,5 @@
 // DO NOT EDIT.
+// swift-format-ignore-file
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: healthdata.proto
@@ -12,7 +13,7 @@ import SwiftProtobuf
 // If the compiler emits an error on this type, it is because this file
 // was generated by a version of the `protoc` Swift plug-in that is
 // incompatible with the version of SwiftProtobuf to which you are linking.
-// Please ensure that your are building against the same version of the API
+// Please ensure that you are building against the same version of the API
 // that was used to generate this file.
 fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAPIVersionCheck {
   struct _2: SwiftProtobuf.ProtobufAPIVersion_2 {}
@@ -785,15 +786,42 @@ struct HealthData {
 
   #if !swift(>=4.1)
     static func ==(lhs: HealthData.OneOf_SpecificData, rhs: HealthData.OneOf_SpecificData) -> Bool {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch (lhs, rhs) {
-      case (.emptyData(let l), .emptyData(let r)): return l == r
-      case (.quantityData(let l), .quantityData(let r)): return l == r
-      case (.categoryData(let l), .categoryData(let r)): return l == r
-      case (.workoutData(let l), .workoutData(let r)): return l == r
-      case (.characteristicData(let l), .characteristicData(let r)): return l == r
-      case (.clinicalRecordData(let l), .clinicalRecordData(let r)): return l == r
-      case (.documentData(let l), .documentData(let r)): return l == r
-      case (.correlationData(let l), .correlationData(let r)): return l == r
+      case (.emptyData, .emptyData): return {
+        guard case .emptyData(let l) = lhs, case .emptyData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.quantityData, .quantityData): return {
+        guard case .quantityData(let l) = lhs, case .quantityData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.categoryData, .categoryData): return {
+        guard case .categoryData(let l) = lhs, case .categoryData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.workoutData, .workoutData): return {
+        guard case .workoutData(let l) = lhs, case .workoutData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.characteristicData, .characteristicData): return {
+        guard case .characteristicData(let l) = lhs, case .characteristicData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.clinicalRecordData, .clinicalRecordData): return {
+        guard case .clinicalRecordData(let l) = lhs, case .clinicalRecordData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.documentData, .documentData): return {
+        guard case .documentData(let l) = lhs, case .documentData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
+      case (.correlationData, .correlationData): return {
+        guard case .correlationData(let l) = lhs, case .correlationData(let r) = rhs else { preconditionFailure() }
+        return l == r
+      }()
       default: return false
       }
     }
@@ -1066,28 +1094,29 @@ struct StatisticsData {
     // methods supported on all messages.
 
     var source: SourceRevision {
-      get {return _storage._source ?? SourceRevision()}
-      set {_uniqueStorage()._source = newValue}
+      get {return _source ?? SourceRevision()}
+      set {_source = newValue}
     }
     /// Returns true if `source` has been explicitly set.
-    var hasSource: Bool {return _storage._source != nil}
+    var hasSource: Bool {return self._source != nil}
     /// Clears the value of `source`. Subsequent reads from it will return its default value.
-    mutating func clearSource() {_uniqueStorage()._source = nil}
+    mutating func clearSource() {self._source = nil}
 
     var data: StatisticsData {
-      get {return _storage._data ?? StatisticsData()}
-      set {_uniqueStorage()._data = newValue}
+      get {return _data ?? StatisticsData()}
+      set {_data = newValue}
     }
     /// Returns true if `data` has been explicitly set.
-    var hasData: Bool {return _storage._data != nil}
+    var hasData: Bool {return self._data != nil}
     /// Clears the value of `data`. Subsequent reads from it will return its default value.
-    mutating func clearData() {_uniqueStorage()._data = nil}
+    mutating func clearData() {self._data = nil}
 
     var unknownFields = SwiftProtobuf.UnknownStorage()
 
     init() {}
 
-    fileprivate var _storage = _StorageClass.defaultInstance
+    fileprivate var _source: SourceRevision? = nil
+    fileprivate var _data: StatisticsData? = nil
   }
 
   struct TimeInterval {
@@ -1108,6 +1137,31 @@ struct StatisticsData {
 
   fileprivate var _storage = _StorageClass.defaultInstance
 }
+
+#if swift(>=5.5) && canImport(_Concurrency)
+extension HealthTypes: @unchecked Sendable {}
+extension RequestSorting: @unchecked Sendable {}
+extension StatisticsOptions: @unchecked Sendable {}
+extension HealthTypeList: @unchecked Sendable {}
+extension HealthDataRequest: @unchecked Sendable {}
+extension HealthDataRequestList: @unchecked Sendable {}
+extension HealthData: @unchecked Sendable {}
+extension HealthData.OneOf_SpecificData: @unchecked Sendable {}
+extension HealthDataList: @unchecked Sendable {}
+extension EmptySpecificData: @unchecked Sendable {}
+extension QuantitySpecificData: @unchecked Sendable {}
+extension CategorySpecificData: @unchecked Sendable {}
+extension WorkoutSpecificData: @unchecked Sendable {}
+extension CharacteristicSpecificData: @unchecked Sendable {}
+extension ClinicalRecordSpecificData: @unchecked Sendable {}
+extension DocumentSpecificData: @unchecked Sendable {}
+extension CorrelationSpecificData: @unchecked Sendable {}
+extension SourceRevision: @unchecked Sendable {}
+extension StatisticsRequest: @unchecked Sendable {}
+extension StatisticsData: @unchecked Sendable {}
+extension StatisticsData.StatisticsDataBySource: @unchecked Sendable {}
+extension StatisticsData.TimeInterval: @unchecked Sendable {}
+#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
@@ -1257,8 +1311,11 @@ extension HealthTypeList: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try decoder.decodeRepeatedEnumField(value: &self.types)
+      case 1: try { try decoder.decodeRepeatedEnumField(value: &self.types) }()
       default: break
       }
     }
@@ -1291,13 +1348,16 @@ extension HealthDataRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplem
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try decoder.decodeSingularEnumField(value: &self.type)
-      case 2: try decoder.decodeSingularStringField(value: &self.startDate)
-      case 3: try decoder.decodeSingularStringField(value: &self.endDate)
-      case 4: try decoder.decodeSingularInt32Field(value: &self.limit)
-      case 5: try decoder.decodeSingularEnumField(value: &self.sorting)
-      case 6: try decoder.decodeRepeatedStringField(value: &self.units)
+      case 1: try { try decoder.decodeSingularEnumField(value: &self.type) }()
+      case 2: try { try decoder.decodeSingularStringField(value: &self.startDate) }()
+      case 3: try { try decoder.decodeSingularStringField(value: &self.endDate) }()
+      case 4: try { try decoder.decodeSingularInt32Field(value: &self.limit) }()
+      case 5: try { try decoder.decodeSingularEnumField(value: &self.sorting) }()
+      case 6: try { try decoder.decodeRepeatedStringField(value: &self.units) }()
       default: break
       }
     }
@@ -1345,8 +1405,11 @@ extension HealthDataRequestList: SwiftProtobuf.Message, SwiftProtobuf._MessageIm
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try decoder.decodeRepeatedMessageField(value: &self.requests)
+      case 1: try { try decoder.decodeRepeatedMessageField(value: &self.requests) }()
       default: break
       }
     }
@@ -1426,79 +1489,122 @@ extension HealthData: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementatio
     _ = _uniqueStorage()
     try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
       while let fieldNumber = try decoder.nextFieldNumber() {
+        // The use of inline closures is to circumvent an issue where the compiler
+        // allocates stack space for every case branch when no optimizations are
+        // enabled. https://github.com/apple/swift-protobuf/issues/1034
         switch fieldNumber {
-        case 1: try decoder.decodeSingularEnumField(value: &_storage._type)
-        case 2: try decoder.decodeSingularStringField(value: &_storage._objectType)
-        case 3: try decoder.decodeSingularStringField(value: &_storage._startDate)
-        case 4: try decoder.decodeSingularStringField(value: &_storage._endDate)
-        case 5: try decoder.decodeSingularStringField(value: &_storage._device)
-        case 6: try decoder.decodeSingularStringField(value: &_storage._metadata)
-        case 7: try decoder.decodeSingularStringField(value: &_storage._uuid)
-        case 8: try decoder.decodeSingularMessageField(value: &_storage._source)
-        case 10:
+        case 1: try { try decoder.decodeSingularEnumField(value: &_storage._type) }()
+        case 2: try { try decoder.decodeSingularStringField(value: &_storage._objectType) }()
+        case 3: try { try decoder.decodeSingularStringField(value: &_storage._startDate) }()
+        case 4: try { try decoder.decodeSingularStringField(value: &_storage._endDate) }()
+        case 5: try { try decoder.decodeSingularStringField(value: &_storage._device) }()
+        case 6: try { try decoder.decodeSingularStringField(value: &_storage._metadata) }()
+        case 7: try { try decoder.decodeSingularStringField(value: &_storage._uuid) }()
+        case 8: try { try decoder.decodeSingularMessageField(value: &_storage._source) }()
+        case 10: try {
           var v: EmptySpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .emptyData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .emptyData(v)}
-        case 11:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .emptyData(v)
+          }
+        }()
+        case 11: try {
           var v: QuantitySpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .quantityData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .quantityData(v)}
-        case 12:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .quantityData(v)
+          }
+        }()
+        case 12: try {
           var v: CategorySpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .categoryData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .categoryData(v)}
-        case 13:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .categoryData(v)
+          }
+        }()
+        case 13: try {
           var v: WorkoutSpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .workoutData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .workoutData(v)}
-        case 14:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .workoutData(v)
+          }
+        }()
+        case 14: try {
           var v: CharacteristicSpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .characteristicData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .characteristicData(v)}
-        case 15:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .characteristicData(v)
+          }
+        }()
+        case 15: try {
           var v: ClinicalRecordSpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .clinicalRecordData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .clinicalRecordData(v)}
-        case 16:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .clinicalRecordData(v)
+          }
+        }()
+        case 16: try {
           var v: DocumentSpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .documentData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .documentData(v)}
-        case 17:
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .documentData(v)
+          }
+        }()
+        case 17: try {
           var v: CorrelationSpecificData?
+          var hadOneofValue = false
           if let current = _storage._specificData {
-            try decoder.handleConflictingOneOf()
+            hadOneofValue = true
             if case .correlationData(let m) = current {v = m}
           }
           try decoder.decodeSingularMessageField(value: &v)
-          if let v = v {_storage._specificData = .correlationData(v)}
+          if let v = v {
+            if hadOneofValue {try decoder.handleConflictingOneOf()}
+            _storage._specificData = .correlationData(v)
+          }
+        }()
         default: break
         }
       }
@@ -1507,6 +1613,10 @@ extension HealthData: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementatio
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
     try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every if/case branch local when no optimizations
+      // are enabled. https://github.com/apple/swift-protobuf/issues/1034 and
+      // https://github.com/apple/swift-protobuf/issues/1182
       if _storage._type != .workoutMain {
         try visitor.visitSingularEnumField(value: _storage._type, fieldNumber: 1)
       }
@@ -1528,26 +1638,42 @@ extension HealthData: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementatio
       if !_storage._uuid.isEmpty {
         try visitor.visitSingularStringField(value: _storage._uuid, fieldNumber: 7)
       }
-      if let v = _storage._source {
+      try { if let v = _storage._source {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 8)
-      }
+      } }()
       switch _storage._specificData {
-      case .emptyData(let v)?:
+      case .emptyData?: try {
+        guard case .emptyData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 10)
-      case .quantityData(let v)?:
+      }()
+      case .quantityData?: try {
+        guard case .quantityData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 11)
-      case .categoryData(let v)?:
+      }()
+      case .categoryData?: try {
+        guard case .categoryData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 12)
-      case .workoutData(let v)?:
+      }()
+      case .workoutData?: try {
+        guard case .workoutData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 13)
-      case .characteristicData(let v)?:
+      }()
+      case .characteristicData?: try {
+        guard case .characteristicData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 14)
-      case .clinicalRecordData(let v)?:
+      }()
+      case .clinicalRecordData?: try {
+        guard case .clinicalRecordData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 15)
-      case .documentData(let v)?:
+      }()
+      case .documentData?: try {
+        guard case .documentData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 16)
-      case .correlationData(let v)?:
+      }()
+      case .correlationData?: try {
+        guard case .correlationData(let v)? = _storage._specificData else { preconditionFailure() }
         try visitor.visitSingularMessageField(value: v, fieldNumber: 17)
+      }()
       case nil: break
       }
     }
@@ -1585,8 +1711,11 @@ extension HealthDataList: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try decoder.decodeRepeatedMessageField(value: &self.data)
+      case 1: try { try decoder.decodeRepeatedMessageField(value: &self.data) }()
       default: break
       }
     }
@@ -1635,10 +1764,13 @@ extension QuantitySpecificData: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularInt64Field(value: &self.count)
-      case 102: try decoder.decodeSingularStringField(value: &self.quantityUnit)
-      case 103: try decoder.decodeSingularDoubleField(value: &self.quantity)
+      case 101: try { try decoder.decodeSingularInt64Field(value: &self.count) }()
+      case 102: try { try decoder.decodeSingularStringField(value: &self.quantityUnit) }()
+      case 103: try { try decoder.decodeSingularDoubleField(value: &self.quantity) }()
       default: break
       }
     }
@@ -1674,8 +1806,11 @@ extension CategorySpecificData: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularInt64Field(value: &self.value)
+      case 101: try { try decoder.decodeSingularInt64Field(value: &self.value) }()
       default: break
       }
     }
@@ -1707,12 +1842,15 @@ extension WorkoutSpecificData: SwiftProtobuf.Message, SwiftProtobuf._MessageImpl
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularDoubleField(value: &self.totalEnergyBurned)
-      case 102: try decoder.decodeSingularStringField(value: &self.totalEnergyBurnedUnit)
-      case 103: try decoder.decodeSingularDoubleField(value: &self.totalDistance)
-      case 104: try decoder.decodeSingularStringField(value: &self.totalDistanceUnit)
-      case 105: try decoder.decodeSingularDoubleField(value: &self.duration)
+      case 101: try { try decoder.decodeSingularDoubleField(value: &self.totalEnergyBurned) }()
+      case 102: try { try decoder.decodeSingularStringField(value: &self.totalEnergyBurnedUnit) }()
+      case 103: try { try decoder.decodeSingularDoubleField(value: &self.totalDistance) }()
+      case 104: try { try decoder.decodeSingularStringField(value: &self.totalDistanceUnit) }()
+      case 105: try { try decoder.decodeSingularDoubleField(value: &self.duration) }()
       default: break
       }
     }
@@ -1756,8 +1894,11 @@ extension CharacteristicSpecificData: SwiftProtobuf.Message, SwiftProtobuf._Mess
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularStringField(value: &self.value)
+      case 101: try { try decoder.decodeSingularStringField(value: &self.value) }()
       default: break
       }
     }
@@ -1786,9 +1927,12 @@ extension ClinicalRecordSpecificData: SwiftProtobuf.Message, SwiftProtobuf._Mess
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularStringField(value: &self.displayName)
-      case 102: try decoder.decodeSingularStringField(value: &self.fhirResource)
+      case 101: try { try decoder.decodeSingularStringField(value: &self.displayName) }()
+      case 102: try { try decoder.decodeSingularStringField(value: &self.fhirResource) }()
       default: break
       }
     }
@@ -1824,12 +1968,15 @@ extension DocumentSpecificData: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularStringField(value: &self.authorName)
-      case 102: try decoder.decodeSingularStringField(value: &self.custodianName)
-      case 103: try decoder.decodeSingularStringField(value: &self.documentData)
-      case 104: try decoder.decodeSingularStringField(value: &self.patientName)
-      case 105: try decoder.decodeSingularStringField(value: &self.title)
+      case 101: try { try decoder.decodeSingularStringField(value: &self.authorName) }()
+      case 102: try { try decoder.decodeSingularStringField(value: &self.custodianName) }()
+      case 103: try { try decoder.decodeSingularStringField(value: &self.documentData) }()
+      case 104: try { try decoder.decodeSingularStringField(value: &self.patientName) }()
+      case 105: try { try decoder.decodeSingularStringField(value: &self.title) }()
       default: break
       }
     }
@@ -1873,8 +2020,11 @@ extension CorrelationSpecificData: SwiftProtobuf.Message, SwiftProtobuf._Message
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeRepeatedMessageField(value: &self.objects)
+      case 101: try { try decoder.decodeRepeatedMessageField(value: &self.objects) }()
       default: break
       }
     }
@@ -1906,12 +2056,15 @@ extension SourceRevision: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 101: try decoder.decodeSingularStringField(value: &self.version)
-      case 102: try decoder.decodeSingularStringField(value: &self.operatingSystemVersion)
-      case 103: try decoder.decodeSingularStringField(value: &self.productType)
-      case 104: try decoder.decodeSingularStringField(value: &self.name)
-      case 105: try decoder.decodeSingularStringField(value: &self.bundleIdentifier)
+      case 101: try { try decoder.decodeSingularStringField(value: &self.version) }()
+      case 102: try { try decoder.decodeSingularStringField(value: &self.operatingSystemVersion) }()
+      case 103: try { try decoder.decodeSingularStringField(value: &self.productType) }()
+      case 104: try { try decoder.decodeSingularStringField(value: &self.name) }()
+      case 105: try { try decoder.decodeSingularStringField(value: &self.bundleIdentifier) }()
       default: break
       }
     }
@@ -1958,11 +2111,14 @@ extension StatisticsRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplem
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try decoder.decodeSingularEnumField(value: &self.type)
-      case 2: try decoder.decodeSingularStringField(value: &self.startDate)
-      case 3: try decoder.decodeSingularStringField(value: &self.endDate)
-      case 4: try decoder.decodeRepeatedEnumField(value: &self.options)
+      case 1: try { try decoder.decodeSingularEnumField(value: &self.type) }()
+      case 2: try { try decoder.decodeSingularStringField(value: &self.startDate) }()
+      case 3: try { try decoder.decodeSingularStringField(value: &self.endDate) }()
+      case 4: try { try decoder.decodeRepeatedEnumField(value: &self.options) }()
       default: break
       }
     }
@@ -2050,17 +2206,20 @@ extension StatisticsData: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
     _ = _uniqueStorage()
     try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
       while let fieldNumber = try decoder.nextFieldNumber() {
+        // The use of inline closures is to circumvent an issue where the compiler
+        // allocates stack space for every case branch when no optimizations are
+        // enabled. https://github.com/apple/swift-protobuf/issues/1034
         switch fieldNumber {
-        case 1: try decoder.decodeRepeatedMessageField(value: &_storage._sources)
-        case 2: try decoder.decodeSingularMessageField(value: &_storage._averageQuantity)
-        case 3: try decoder.decodeSingularMessageField(value: &_storage._maximumQuantity)
-        case 4: try decoder.decodeSingularMessageField(value: &_storage._minimumQuantity)
-        case 5: try decoder.decodeSingularMessageField(value: &_storage._sumQuantity)
-        case 6: try decoder.decodeSingularMessageField(value: &_storage._duration)
-        case 7: try decoder.decodeSingularMessageField(value: &_storage._mostRecentQuantity)
-        case 8: try decoder.decodeSingularMessageField(value: &_storage._mostRecentQuantityDateInterval)
-        case 9: try decoder.decodeSingularMessageField(value: &_storage._dataInterval)
-        case 10: try decoder.decodeRepeatedMessageField(value: &_storage._dataBySource)
+        case 1: try { try decoder.decodeRepeatedMessageField(value: &_storage._sources) }()
+        case 2: try { try decoder.decodeSingularMessageField(value: &_storage._averageQuantity) }()
+        case 3: try { try decoder.decodeSingularMessageField(value: &_storage._maximumQuantity) }()
+        case 4: try { try decoder.decodeSingularMessageField(value: &_storage._minimumQuantity) }()
+        case 5: try { try decoder.decodeSingularMessageField(value: &_storage._sumQuantity) }()
+        case 6: try { try decoder.decodeSingularMessageField(value: &_storage._duration) }()
+        case 7: try { try decoder.decodeSingularMessageField(value: &_storage._mostRecentQuantity) }()
+        case 8: try { try decoder.decodeSingularMessageField(value: &_storage._mostRecentQuantityDateInterval) }()
+        case 9: try { try decoder.decodeSingularMessageField(value: &_storage._dataInterval) }()
+        case 10: try { try decoder.decodeRepeatedMessageField(value: &_storage._dataBySource) }()
         default: break
         }
       }
@@ -2069,33 +2228,37 @@ extension StatisticsData: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
     try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every if/case branch local when no optimizations
+      // are enabled. https://github.com/apple/swift-protobuf/issues/1034 and
+      // https://github.com/apple/swift-protobuf/issues/1182
       if !_storage._sources.isEmpty {
         try visitor.visitRepeatedMessageField(value: _storage._sources, fieldNumber: 1)
       }
-      if let v = _storage._averageQuantity {
+      try { if let v = _storage._averageQuantity {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
-      }
-      if let v = _storage._maximumQuantity {
+      } }()
+      try { if let v = _storage._maximumQuantity {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
-      }
-      if let v = _storage._minimumQuantity {
+      } }()
+      try { if let v = _storage._minimumQuantity {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 4)
-      }
-      if let v = _storage._sumQuantity {
+      } }()
+      try { if let v = _storage._sumQuantity {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 5)
-      }
-      if let v = _storage._duration {
+      } }()
+      try { if let v = _storage._duration {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 6)
-      }
-      if let v = _storage._mostRecentQuantity {
+      } }()
+      try { if let v = _storage._mostRecentQuantity {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 7)
-      }
-      if let v = _storage._mostRecentQuantityDateInterval {
+      } }()
+      try { if let v = _storage._mostRecentQuantityDateInterval {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 8)
-      }
-      if let v = _storage._dataInterval {
+      } }()
+      try { if let v = _storage._dataInterval {
         try visitor.visitSingularMessageField(value: v, fieldNumber: 9)
-      }
+      } }()
       if !_storage._dataBySource.isEmpty {
         try visitor.visitRepeatedMessageField(value: _storage._dataBySource, fieldNumber: 10)
       }
@@ -2134,63 +2297,36 @@ extension StatisticsData.StatisticsDataBySource: SwiftProtobuf.Message, SwiftPro
     2: .same(proto: "data"),
   ]
 
-  fileprivate class _StorageClass {
-    var _source: SourceRevision? = nil
-    var _data: StatisticsData? = nil
-
-    static let defaultInstance = _StorageClass()
-
-    private init() {}
-
-    init(copying source: _StorageClass) {
-      _source = source._source
-      _data = source._data
-    }
-  }
-
-  fileprivate mutating func _uniqueStorage() -> _StorageClass {
-    if !isKnownUniquelyReferenced(&_storage) {
-      _storage = _StorageClass(copying: _storage)
-    }
-    return _storage
-  }
-
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    _ = _uniqueStorage()
-    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
-      while let fieldNumber = try decoder.nextFieldNumber() {
-        switch fieldNumber {
-        case 1: try decoder.decodeSingularMessageField(value: &_storage._source)
-        case 2: try decoder.decodeSingularMessageField(value: &_storage._data)
-        default: break
-        }
+    while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
+      switch fieldNumber {
+      case 1: try { try decoder.decodeSingularMessageField(value: &self._source) }()
+      case 2: try { try decoder.decodeSingularMessageField(value: &self._data) }()
+      default: break
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
-      if let v = _storage._source {
-        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-      }
-      if let v = _storage._data {
-        try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
-      }
-    }
+    // The use of inline closures is to circumvent an issue where the compiler
+    // allocates stack space for every if/case branch local when no optimizations
+    // are enabled. https://github.com/apple/swift-protobuf/issues/1034 and
+    // https://github.com/apple/swift-protobuf/issues/1182
+    try { if let v = self._source {
+      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+    } }()
+    try { if let v = self._data {
+      try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+    } }()
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: StatisticsData.StatisticsDataBySource, rhs: StatisticsData.StatisticsDataBySource) -> Bool {
-    if lhs._storage !== rhs._storage {
-      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
-        let _storage = _args.0
-        let rhs_storage = _args.1
-        if _storage._source != rhs_storage._source {return false}
-        if _storage._data != rhs_storage._data {return false}
-        return true
-      }
-      if !storagesAreEqual {return false}
-    }
+    if lhs._source != rhs._source {return false}
+    if lhs._data != rhs._data {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -2205,9 +2341,12 @@ extension StatisticsData.TimeInterval: SwiftProtobuf.Message, SwiftProtobuf._Mes
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 3: try decoder.decodeSingularStringField(value: &self.startDate)
-      case 4: try decoder.decodeSingularStringField(value: &self.endDate)
+      case 3: try { try decoder.decodeSingularStringField(value: &self.startDate) }()
+      case 4: try { try decoder.decodeSingularStringField(value: &self.endDate) }()
       default: break
       }
     }
